chore: added pattern for input
