+++ apps/database-new/app/[threadId]/Chat.tsx
@@ -3,14 +3,13 @@ import { useMutation } from '@tanstack/react-query'
 import { sortBy } from 'lodash'
 import { Loader2 } from 'lucide-react'
 import { useParams, useRouter } from 'next/navigation'
-import { useEffect, useMemo, useState } from 'react'
-import { Input, ScrollArea, cn } from 'ui'
+import { createRef, useEffect, useMemo, useState } from 'react'
+import { AssistantChatInput, Input, ScrollArea, cn } from 'ui'
 
 import { useMessagesQuery } from '@/data/messages-query'
 import { AssistantMessage, UserMessage } from '@/lib/types'
 import BottomMarker from './BottomMarker'
 import UserChat from './UserChat'
-import { ChatInputAtom } from '@/components/Chat/ChatInput'
 
 export const Chat = () => {
   const router = useRouter()
@@ -57,6 +56,19 @@ export const Chat = () => {
     setValue('')
   }, [loading])
 
+  const textAreaRef = createRef<HTMLTextAreaElement>()
+  const submitRef = createRef<HTMLButtonElement>()
+
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    // Check if the pressed key is "Enter" (key code 13) without the "Shift" key
+    if (event.key === 'Enter' && !event.shiftKey) {
+      event.preventDefault()
+      if (submitRef.current) {
+        submitRef.current.click()
+      }
+    }
+  }
+
   return (
     <div
       className={cn(
@@ -96,23 +108,29 @@ export const Chat = () => {
         )}
 
         <div className="px-4 pb-4">
-          <ChatInputAtom
-            value={value}
-            disabled={loading || inputEntered}
-            loading={loading || inputEntered}
-            placeholder={
-              loading
-                ? 'Generating reply to request...'
-                : 'Ask for some changes on the selected message'
-            }
-            onChange={(v) => setValue(v.target.value)}
-            handleSubmit={() => {
-              if (value && value.length > 0) {
-                mutate(value)
-                setInputEntered(true)
-              }
+          <form
+            key={`chat-thread-form-${runId}`}
+            id={`chat-thread-form-${runId}`}
+            onSubmit={async (event) => {
+              event.preventDefault()
+              mutate(value)
             }}
-          />
+          >
+            <AssistantChatInput
+              ref={textAreaRef}
+              submitRef={submitRef}
+              value={value}
+              disabled={loading || inputEntered}
+              loading={loading || inputEntered}
+              placeholder={
+                loading
+                  ? 'Generating reply to request...'
+                  : 'Ask for some changes on the selected message'
+              }
+              onChange={(v) => setValue(v.target.value)}
+              onKeyDown={handleKeyDown}
+            />
+          </form>
         </div>
       </div>
     </div>
